+++ src/org/lockss/plugin/highwire/HighWireArchivalUnit.java
@@ -1,5 +1,5 @@
 /*
- * $Id: HighWireArchivalUnit.java,v 1.10 2003-03-08 01:36:44 troberts Exp $
+ * $Id: HighWireArchivalUnit.java,v 1.11 2003-03-08 01:44:32 troberts Exp $
  */
 
 /*
@@ -55,23 +55,23 @@ public class HighWireArchivalUnit extends BaseArchivalUnit {
    */
   public static final String PARAM_HIGHWIRE_PAUSE_TIME =
       Configuration.PREFIX + "highwire.pause.time";
-  private static final int DEFAULT_PAUSE_TIME = 10000;
+  private static final long DEFAULT_PAUSE_TIME = 10000;
 
   private static final String EXPECTED_URL_PATH = "/";
 
   protected Logger logger = Logger.getLogger(LOG_NAME);
-  private int pauseMS;
+  private long pauseMS;
 
   private int volume;
   private URL base;
 
   public static final String PARAM_HIGHWIRE_NC_INTERVAL =
       Configuration.PREFIX + "highwire.nc_interval";
-  private static final int DEFAULT_NC_INTERVAL = 14;
+  private static final long DEFAULT_NC_INTERVAL = 14;
 
   private static final long DAY_MS = 1000 * 60 * 60 * 24;
 
-  private int ncCrawlInterval;
+  private long ncCrawlInterval;
 
 
 
@@ -183,7 +183,7 @@ public class HighWireArchivalUnit extends BaseArchivalUnit {
   }
 
   public void pause() {
-    pause(pauseMS);
+    pause((int)pauseMS);
   }
 
   public String getAUId() {
